bot_2/requirements.txt
python-telegram-bot
aiohttp
aiomysql
asyncio